// src/interfaces/IMail.ts
export interface MessageType {
  appHost?: string
  logo?: string
  headerName?: string
  visitorName?: string
  showroomName?: string
  enterTime?: string
  virtualSpaceName?: string
  virtualSpaceType?: string
  repName?: string
  organizationName?: string
  url?: string
  thumbnail?: string
  brandColor?: string
  expiresDate?: string
  message?: string
  qrCode?: string
  timePeriod?: string
  leaveTime?: string
}